test scripts

+++ test/qtest3.py
@@ -17,7 +17,7 @@ for i in range(num_gpus):
 
 before_mem = torch.cuda.memory_allocated()
 print(f"getting model from {model_folder}")
-model = load_model(model_name, device_map="auto", storage_path=model_folder, fully_parallel=True, quantization="int8")
+model = load_model(model_name, device_map="auto", storage_path=model_folder, fully_parallel=True, quantization="int4")
 after_mem = torch.cuda.memory_allocated()
 print(f"Memory difference: {after_mem - before_mem}")
 print(f"memory footprint: {model.get_memory_footprint()}")
